Fix bug in `gmail.GetMessages.render_message()`
In the case where a GMAIL message has multiple "parts" (i.e., HTML,
images, etc.), the payload['body'] dictionary element is empty, and
what would be there is as a list in the 'parts' element.  In the simple
case where the GMAIL message is text only, there is only one part (as
is the case with relay DOL messages -- the test case when developing
this class).

This commit adds a check for the 'parts' dictionary element, and selects
the first PART, if such a list exists.  Otherwise, the usual behavior of
extracting the payload['body'] dictionary element holds.  This works
with the hand-delivered DOL messages of the past few days (the mail
relay is down due to IT issues), but no guarantee is made or implied
that this fix will be appropriate or work in all use cases of this
j5 class.

	modified:   johnnyfive/gmail.py